package_base: adopt WindowsSimulatedRPath
Allows for healthier references to PackageBase from the simulated rpath class
